Moving code to reports folder
