+++ lib/challenge_gov/reports/accounts_recertified_date_range.ex
@@ -1,8 +1,9 @@
-defmodule ChallengeGov.NumberOfSubmissions do
+defmodule ChallengeGov.Reports.AccountsRecertifiedDateRange do
   @moduledoc false
   import Ecto.Query
 
-  alias ChallengeGov.Challenges.Challenge
+  alias ChallengeGov.Accounts.User
+  alias ChallengeGov.CertificationLogs.CertificationLog
 
   def execute(params) do
     %{
@@ -24,28 +25,22 @@ defmodule ChallengeGov.NumberOfSubmissions do
       |> List.to_tuple()
       |> Timex.to_datetime()
 
-    from(c in Challenge)
-    |> join(:left, [c], a in assoc(c, :agency))
-    |> join(:left, [c, a], s in assoc(c, :submissions))
+    from(u in User)
+    |> join(:left, [u], s in CertificationLog, on: u.id == s.user_id)
     |> where(
-      [c, a, s],
-      fragment("? BETWEEN ? AND ?", c.inserted_at, ^s_date, ^e_date)
+      [u, s],
+      fragment("? BETWEEN ? AND ?", s.certified_at, ^s_date, ^e_date)
     )
-    |> where([c], c.status == "published")
-    |> select([c, a, s], %{
-      challenge_id: c.id,
-      challenge_name: c.title,
-      created_date: c.inserted_at,
+    |> select([u, s], %{
+      user_id: u.id,
+      account_type: u.role,
+      action: "recertified",
+      logged_date: s.certified_at,
+      account_status: u.status,
+      last_login: u.last_active,
       start_date: ^start_date,
-      end_date: ^end_date,
-      listing_type: 'Full',
-      submissions_count: count(s)
+      end_date: ^end_date
     })
-    |> group_by([c, a, s], [
-      c.id,
-      c.title,
-      c.inserted_at
-    ])
     |> ChallengeGov.Repo.all()
     |> build_data_structure()
   end
@@ -57,14 +52,15 @@ defmodule ChallengeGov.NumberOfSubmissions do
 
     Enum.map(active_published_challenge_data, fn c ->
       %{
-        challenge_id: c.challenge_id,
-        challenge_name: c.challenge_name,
+        user_id: c.user_id,
+        account_type: c.account_type,
+        action: c.action,
+        logged_date: c.logged_date,
+        account_status: c.account_status,
+        last_login: c.last_login,
         start_date: c.start_date,
         end_date: c.end_date,
-        created_date: c.created_date,
-        submissions: c.submissions_count,
-        current_timestamp: now,
-        listing_type: c.listing_type
+        current_timestamp: now
       }
     end)
   end
